update fmt4_csv.py
The purpose is for converting fmt4 file to csv file to get a direct look at the data. 
In this one, all variables not closely related to lactation curve are commented out, first 30000 cow records which born after 2008 is converted, test day data are in JSON file format in the last column of the output. The main lines of the code could also be used for python analysis directly from fmt4 file.